Create README.md
# JAVA SCRIPT #
**JavaScript, often abbreviated as JS, is a high-level, dynamic, weakly typed, object-based,multi-paradigm, and interpreted 
client-side programming language. Alongside HTML and CSS, JavaScript is one of the three core technologies of World Wide Web 
content production.It is used to make webpages interactive and provide online programs, including video games.**

## area.js##
In geometry, Heron's formula named after Hero of Alexandria, gives the area of a triangle by requiring no arbitrary choice of 
side as base or vertex as origin, contrary to other formulas for the area of a triangle, such as half the base times the height
or half the norm of a cross product of two sides.

*Area of a triangle*
The *Math.sqrt()* function is used to get the square root of a number. If the value of the number is negative, *Math.sqrt()* 
returns NaN.

## day.js ##
The *getDay()* method is used to get day of the week for the specified date according to local time, where 0 represents Sunday.
The value returned by getDay() is an integer corresponding to the day of the week: 0 for Sunday, 1 for Monday, 2 for Tuesday, 
and so on.

## index.js ##
* The *toString()* method converts a number to a string.
* The *charAt()* method returns the character at the specified index in a string.
  The index of the first character is 0, the second character is 1, and so on.
  
## today.js ##
* The *getDate()* method is used to get the day of the month for the specified date according to local time. The value returned 
by getDate() is an integer between 1 and 31.

* The *getMonth()* method returns the month in the specified date according to local time, as a zero-based value (where zero 
indicates the first month of the year). The value returned by getMonth() is an integer between 0 and 11. 0 corresponds to 
January, 1 to February, and so on.

* The *getFullYear()* method is used to get the year of the specified date according to local time. The value returned by the 
method is an absolute number. For dates between the years 1000 and 9999, getFullYear() returns a four-digit number, for 
example, 1985.